refactor: structure files
